Refactor: 서버 코드수정
