--- package.json
@@ -1,29 +1,0 @@
-{
-  "name": "music-flower",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node app.js",
-    "test": "node test.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/codestates/Music-Flower.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/codestates/Music-Flower/issues"
-  },
-  "homepage": "https://github.com/codestates/Music-Flower#readme",
-  "dependencies": {
-    "react": "^17.0.2",
-    "sequelize": "^6.17.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.17.5",
-    "sequelize-cli": "^6.4.1"
-  }
-}
