+++ server/package.json
@@ -105,6 +105,7 @@
     "jsesc": "^2.5.2",
     "json-buffer": "^3.0.0",
     "json5": "^2.2.0",
+    "jsonwebtoken": "^8.5.1",
     "keyv": "^3.1.0",
     "latest-version": "^5.1.0",
     "loose-envify": "^1.4.0",
